requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.12.4
django-filter==21.1
django-cors-headers>=3.11.0
